+++ models/__init__.py
@@ -2,6 +2,7 @@
 Module to hold models
 """
 
-__all__ = ["cifar10_loader"]
+__all__ = ["cifar10_loader", "MLP"]
 
+from models.mlp import MLP
 from models.utils import cifar10_loader
