--- permuter/core/__init__.py
@@ -1,11 +1,0 @@
-__all__ = [
-    "get_losses",
-    "combine_models",
-    "permute_model",
-    "ActMatching",
-    "WeightMatching",
-    "STEstimator",
-]
-
-from .algo import ActMatching, STEstimator, WeightMatching
-from .utils import combine_models, get_losses, permute_model
