+++ permuter/mlp.py
@@ -1,27 +1,121 @@
+import copy
 from typing import Optional
 
 import numpy as np
 import torch
+from torch.utils.data import DataLoader
 
-from config import (
-    ACT_PERM,
-    DEVICE,
-    LAMBDA_ARRAY,
-    MLP_MODEL1_PATH,
-    MLP_MODEL2_PATH,
-    WEIGHT_PERM,
-)
-from .core import (
-    ActMatching,
-    STEstimator,
-    WeightMatching,
-    combine_models,
-    get_losses,
-    permute_model,
-)
+from config import (ACT_PERM, BIAS, DEVICE, LAMBDA_ARRAY, MLP_MODEL1_PATH,
+                    MLP_MODEL2_PATH, WEIGHT, WEIGHT_PERM)
 from helper import plt_dict, read_file, write_file
-from models.mlp import MLP, register_hook
-from models import cifar10_loader
+from models import MLP, cifar10_loader
+from models.mlp import register_hook
+from permuter._algo import ActMatching, STEstimator, WeightMatching
+
+
+def permute_model(
+    model: torch.nn.Module,
+    perm_dict: dict[str, torch.Tensor],
+) -> torch.nn.Module:
+    """
+    Permute the model with the dictionary
+
+    :param model: Model to be permuted
+    :type model: torch.nn.Module
+    :param perm_dict: Permutation dictionary
+    :type perm_dict: dict[str, torch.Tensor]
+    :return: Permuted model
+    :rtype: torch.nn.Module
+    """
+    # Creating model instance to hold the permuted model
+    permuted_model = MLP().to(DEVICE)
+
+    perm_state_dict = permuted_model.state_dict()
+    model2_state_dict = model.state_dict()
+
+    for key in perm_state_dict.keys():
+        layer_name, weight_type = key.split(".")
+
+        if weight_type == WEIGHT and not layer_name.endswith("1"):
+            _layer_name, _layer_num = layer_name.split("_")
+            prev_layer_name = "_".join([_layer_name, str(int(_layer_num) - 1)])
+
+            # Considers both column and row permutation if applicable else only column transformation
+            # The latter case happens for last layer
+            perm_state_dict[key] = (
+                (
+                    perm_dict[layer_name]
+                    @ model2_state_dict[key]
+                    @ perm_dict[prev_layer_name].T
+                )
+                if layer_name in perm_dict
+                else model2_state_dict[key] @ perm_dict[prev_layer_name].T
+            )
+        elif layer_name in perm_dict:
+            perm_state_dict[key] = perm_dict[layer_name] @ model2_state_dict[key]
+
+    permuted_model.load_state_dict(perm_state_dict)
+    return permuted_model
+
+
+def combine_models(
+    model1: torch.nn.Module, model2: torch.nn.Module, lam: float
+) -> torch.nn.Module:
+    """
+    Combine models using linear interpolation (1-lam)*model1 + lam*model2
+
+    :param model1: Model 1
+    :type model1: torch.nn.Module
+    :param model2: Model 2
+    :type model2: torch.nn.Module
+    :param lam: Lambda value in linear interpolation way
+    :type lam: float
+    :return: Combined model
+    :rtype: torch.nn.Module
+    """
+    # Creating dummy model
+    model3 = copy.deepcopy(model1).to(DEVICE)
+    model3_state_dict = model3.state_dict()
+    model1.to(DEVICE)
+    model2.to(DEVICE)
+    state_dict1 = model1.state_dict()
+    state_dict2 = model2.state_dict()
+
+    for key in model3_state_dict.keys():
+        model3_state_dict[key] = (1 - lam) * state_dict1[key] + lam * state_dict2[key]
+
+    model3.load_state_dict(model3_state_dict)
+    return model3
+
+
+def get_losses(
+    data_loader: DataLoader,
+    combined_models: list[torch.nn.Module],
+) -> np.ndarray:
+    """
+    Generates data for loss barrier plot
+
+    :param data_loader: Data Loader
+    :type data_loader: DataLoader
+    :param model1: Model 1
+    :type model1: torch.nn.Module
+    :param model2: Model 2
+    :type model2: torch.nn.Module
+    :param combined_models: list of combined models for different values of lambda
+    :type combined_models: list[torch.nn.Module]
+    :return: Loss barrier for combined models
+    :rtype: np.ndarray
+    """
+    loss = [0.0 for _ in range(len(combined_models))]
+    for inp, out in data_loader:
+        for ix, model in enumerate(combined_models):
+            loss[ix] += torch.nn.functional.cross_entropy(
+                model(inp.to(DEVICE)), out.to(DEVICE), reduction="sum"
+            ).item()
+
+    return np.array(loss) / len(data_loader.dataset)  # type: ignore
+
+
 
 def activation_matching() -> dict[str, torch.Tensor]:
     """
@@ -161,7 +255,7 @@ def generate_plots(
     return result
 
 
-if __name__ == "__main__":
+def run():
 
     # if not WEIGHT_PERM.is_file():
     #     weight_perm = weight_matching()
@@ -199,7 +293,7 @@ if __name__ == "__main__":
 
     ste = STEstimator(arch=[512, 512, 512, 10])
     perm, losses = ste.evaluate_permutation(
-        model1=mlp_model1, model2=mlp_model2, data_loader=train_loader
+        model1=mlp_model1, model2=mlp_model2, data_loader=train_loader, permute_model= permute_model
     )
 
     results_dict = generate_plots(model1=mlp_model1, model2=mlp_model2, ste_perm=perm)
